% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.cindex.R
\name{cal.cindex}
\alias{cal.cindex}
\title{compute C index for a Cox model}
\usage{
cal.cindex(pred, y, weights = rep(1, nrow(y)))
}
\arguments{
\item{pred}{Predictions from a \code{"cox"} object}

\item{y}{a survival response object - a matrix with two columns "time" and
"status"; see documentation for "glmnet" or see documentation for "GAGA"}

\item{weights}{optional observation weights}
}
\description{
Computes Harrel's C index for predictions from a \code{"cox"} object.
}
\details{
Computes the concordance index, taking into account censoring.
This file fully references the Cindex.R file in glmnet package.
}
\examples{

set.seed(10101)
N = 1000
p = 30
nzc = p/3
x = matrix(rnorm(N * p), N, p)
beta = rnorm(nzc)
fx = x[, seq(nzc)] \%*\% beta/3
hx = exp(fx)
ty = rexp(N, hx)
tcens = rbinom(n = N, prob = 0.3, size = 1)  # censoring indicator
y = cbind(time = ty, status = 1 - tcens)  # y=Surv(ty,1-tcens) with library(survival)
fit = GAGA(x, y, family = "cox")
pred = predict.GAGA(fit, newx = x)
cat("\n Cindex:", cal.cindex(pred, y))


}
\references{
Harrel Jr, F. E. and Lee, K. L. and Mark, D. B. (1996)
\emph{Tutorial in biostatistics: multivariable prognostic models: issues in
developing models, evaluating assumptions and adequacy, and measuring and
reducing error}, Statistics in Medicine, 15, pages 361--387.#'
}
\seealso{
\code{cv.glmnet}
}
\author{
Trevor Hastie  <hastie@stanford.edu>
}
