% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAGA.R
\name{GAGA}
\alias{GAGA}
\title{Title}
\usage{
GAGA(
  X,
  y,
  family = c("gaussian", "binomial", "poisson", "multinomial", "cox"),
  alpha = 2,
  itrNum = 30,
  QR_flag = FALSE,
  flag = TRUE,
  lamda_0 = 0.001,
  fdiag = FALSE
)
}
\arguments{
\item{fdiag}{}
}
\description{
Title
}
