\name{gagaReport}
\alias{gagaReport}
\title{View and create output of GAGA objects}
\usage{
gagaReport(gagaInput, gagaOutput, outType = "complete", yRange = c(-250, 0),
  output_file_prefix = "")
}
\arguments{
  \item{gagaInput}{Raw data put into gaga()}

  \item{gagaOutput}{Object of class ga produced by gaga()}

  \item{outType}{Type of output desired}

  \item{yRange}{Y-axis range when plotting fitness of
  individuals}

  \item{output_file_prefix}{Optional prefix for output
  files}
}
\value{
Description of the returned object
}
\description{
Description here
}
\examples{
gagaReport(blah,UNKNOWN)
}
\author{
Alex Murison and Christopher Wardell
\email{Alexander.Murison@icr.ac.uk}
}

