% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LM_GAGA.R
\name{LM_GAGA}
\alias{LM_GAGA}
\title{Title}
\usage{
LM_GAGA(
  X,
  y,
  alpha = 3,
  itrNum = 50,
  QR_flag = FALSE,
  flag = TRUE,
  lamda_0 = 0.001,
  fix_sigma = FALSE,
  sigm2_0 = 1
)
}
\arguments{
\item{sigm2_0}{}
}
\description{
Title
}
