% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_GAGA.R
\name{multinomial_GAGA}
\alias{multinomial_GAGA}
\title{Title}
\usage{
multinomial_GAGA(
  X,
  y,
  alpha = 2,
  itrNum = 50,
  flag = TRUE,
  lamda_0 = 0.001,
  fdiag = TRUE
)
}
\arguments{
\item{X}{design matrix}

\item{fdiag}{}
}
\description{
Title
}
