% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_GAGA.R
\name{logistic_GAGA}
\alias{logistic_GAGA}
\title{Title}
\usage{
logistic_GAGA(
  X,
  y,
  alpha = 1,
  itrNum = 50,
  flag = TRUE,
  lamda_0 = 0.001,
  fdiag = TRUE
)
}
\arguments{
\item{fdiag}{}
}
\description{
Title
}
