% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_GAGA.R
\name{poisson_GAGA}
\alias{poisson_GAGA}
\title{Title}
\usage{
poisson_GAGA(
  X,
  y,
  alpha = 1,
  itrNum = 30,
  flag = TRUE,
  lamda_0 = 0.5,
  fdiag = TRUE
)
}
\arguments{
\item{fdiag}{}
}
\description{
Title
}
